requirements.txt
anyio==3.6.2
asyncio==3.4.3
boto3==1.26.99
botocore==1.29.99
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
fastapi==0.92.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
jmespath==1.0.1
Pillow==9.4.0
pip==23.2.1
pydantic==1.10.5
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
s3transfer==0.6.0
setuptools==67.3.3
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
wheel==0.38.4
python-jose==3.3.0
passlib==1.7.4